BarTender_Bot 0.0.4/pyproject.toml
[tool.poetry]
name = "barpatiot1-bot-ver 0-03"
version = "0.1.0"
description = ""
authors = ["TTwenzos"]
readme = "README.md"
packages = [{include = "barpatiot1_bot_ver 0.3"}]

[tool.poetry.dependencies]
python = "^3.12"
telebot = "^0.0.5"
openpyxl = "^3.1.5"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
